+++ CHANGELOG.md
@@ -4,6 +4,13 @@
 
 - Add SentryObjC wrapper SDK to provide pure Objective-C compatibility layer that wraps the main Sentry SDK with stable, documented Objective-C interfaces. (#7598)
   This SDK is designed for projects that require strict Objective-C compatibility without Swift dependencies.
+### Features
+
+- Add `SentrySDK.lastRunStatus` to distinguish unknown, no-crash and crash (#7469)
+
+### Fixes
+
+- Make SentryBreadcrumb thread-safe to prevent crashes in addBreadcrumb ([#7665](https://github.com/getsentry/sentry-cocoa/pull/7665))
 
 ## 9.7.0